moved cell tallies to another file
