+++ openmc_dagmc_wrapper/__init__.py
@@ -3,8 +3,7 @@ from .utils import create_material, diff_between_angles
 from .Geometry import Geometry
 from .Materials import Materials
 from .Tally import (
-    CellTally,
-    CellTallies,
+    Tally,
     MeshTally2D,
     MeshTallies2D,
     MeshTally3D,
@@ -12,4 +11,5 @@ from .Tally import (
     TetMeshTally,
     TetMeshTallies,
 )
+from .tallies import *
 from .Settings import FusionSettings
