+++ openmc_dagmc_wrapper/tallies/__init__.py
@@ -1,0 +1,1 @@
+from .cell_tally import CellTally, CellTallies
